+++ src/builder/mod.rs
@@ -1,11 +1,3 @@
-use crate::ast::Module;
-use crate::generator;
-use crate::lexer;
-use crate::parser;
-use crate::Lib;
-use crate::PathBuf;
-use generator::Generator;
-use std::env;
 /**
  * Copyright 2021 Garrit Franke
  *
@@ -21,6 +13,13 @@ use std::env;
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+use crate::ast::Module;
+use crate::generator::{self, Generator, Target};
+use crate::lexer;
+use crate::parser;
+use crate::Lib;
+use crate::PathBuf;
+use std::env;
 use std::fs::File;
 use std::io::Read;
 use std::io::Write;
@@ -124,8 +123,8 @@ impl Builder {
 
     pub(crate) fn generate(
         &mut self,
-        target: generator::Target,
-        out_file: PathBuf,
+        target: &Target,
+        buffer: &mut Box<impl Write>,
     ) -> Result<(), String> {
         let mut mod_iter = self.modules.iter();
 
@@ -136,9 +135,9 @@ impl Builder {
         }
 
         let output = match target {
-            generator::Target::JS => generator::js::JsGenerator::generate(condensed),
-            generator::Target::C => generator::c::CGenerator::generate(condensed),
-            generator::Target::LLVM => {
+            Target::JS => generator::js::JsGenerator::generate(condensed),
+            Target::C => generator::c::CGenerator::generate(condensed),
+            Target::LLVM => {
                 #[cfg(not(feature = "llvm"))]
                 panic!("'llvm' feature should be enabled to use LLVM target");
 
@@ -147,9 +146,8 @@ impl Builder {
             }
         };
 
-        let mut file = std::fs::File::create(out_file).expect("create failed");
-        file.write_all(output.as_bytes()).expect("write failed");
-        file.flush().map_err(|_| "Could not flush file".into())
+        buffer.write_all(output.as_bytes()).expect("write failed");
+        buffer.flush().map_err(|_| "Could not flush file".into())
     }
 
     fn build_stdlib(&mut self) {