+++ src/command/build.rs
@@ -14,11 +14,27 @@
  * limitations under the License.
  */
 use crate::builder;
-use crate::generator;
+use crate::generator::Target;
+use std::fs::File;
+use std::io::Write;
 use std::path::Path;
 
-pub fn build(target: generator::Target, in_file: &Path, out_file: &Path) -> Result<(), String> {
+pub fn build(target: &Target, in_file: &Path, out_file: &Path) -> Result<(), String> {
+    let mut buf = Box::new(Vec::new());
+    build_to_buffer(target, in_file, &mut buf)?;
+
+    File::create(out_file)
+        .map_err(|e| format!("Could not create output file: {}", e))?
+        .write_all(&buf)
+        .map_err(|e| format!("Could not write to file: {}", e))
+}
+
+pub fn build_to_buffer(
+    target: &Target,
+    in_file: &Path,
+    buf: &mut Box<impl Write>,
+) -> Result<(), String> {
     let mut b = builder::Builder::new(in_file.to_path_buf());
     b.build(&target)?;
-    b.generate(target, out_file.to_path_buf())
+    b.generate(target, buf)
 }