+++ src/command/run.rs
@@ -15,63 +15,40 @@
  */
 use crate::command::build;
 use crate::generator::Target;
+use std::io::Read;
 use std::io::Write;
 use std::path::PathBuf;
-use std::process;
 use std::process::Command;
 use std::process::Stdio;
-use tempfile::tempdir;
 
 pub fn run(target: Target, in_file: PathBuf) -> Result<(), String> {
-    let out_dir = tempdir()
-        .expect("Could not create temporary file")
-        .into_path();
+    let mut buf = Box::new(Vec::new());
+    build::build_to_buffer(&target, &in_file, &mut buf)?;
 
-    let intermediate_out_file_path = out_dir.join("intermediate.c");
-    build::build(target, &in_file, &intermediate_out_file_path)?;
-    let out_file = out_dir.join("out");
     match target {
-        Target::C => {
-            Command::new("/usr/bin/cc")
-                .arg(&intermediate_out_file_path)
-                .arg("-o")
-                .arg(&out_file)
-                .stdout(Stdio::piped())
-                .stderr(Stdio::piped())
-                .output()
-                .expect("Could not spawn compilation process");
-
-            let out = Command::new(out_file)
-                .stdout(Stdio::piped())
-                .stderr(Stdio::piped())
-                .output()
-                .expect("Could not spawn run process");
-
-            std::io::stdout()
-                .write_all(&out.stdout)
-                .expect("Could not write to stdout");
-
-            std::io::stderr()
-                .write_all(&out.stderr)
-                .expect("Could not write to stderr");
-        }
         Target::JS => {
-            let out = Command::new("node")
-                .arg(&intermediate_out_file_path)
+            let process = Command::new("node")
+                .stdin(Stdio::piped())
                 .stdout(Stdio::piped())
                 .stderr(Stdio::piped())
-                .output()
-                .expect("Could not spawn run process");
+                .spawn()
+                .map_err(|e| format!("Could not spawn Node.js process: {}", e))?;
 
+            process
+                .stdin
+                .unwrap()
+                .write_all(&buf)
+                .map_err(|e| format!("Could not write to Node.js process: {}", e))?;
+
+            let mut s = Vec::new();
+            process
+                .stdout
+                .unwrap()
+                .read_to_end(&mut s)
+                .map_err(|e| format!("Could not read from child process: {}", e))?;
             std::io::stdout()
-                .write_all(&out.stdout)
-                .expect("Could not write to stdout");
-
-            std::io::stderr()
-                .write_all(&out.stderr)
-                .expect("Could not write to stderr");
-
-            process::exit(out.status.code().unwrap())
+                .write_all(&s)
+                .map_err(|e| format!("Could not write to stdout: {}", e))?;
         }
         _ => todo!(),
     }